+++ test/unit/fuzz-tools.test.js
@@ -294,6 +294,7 @@ import * as deauvilleV314 from '../../lib/deauville-v314.js';
 import * as jonesV315 from '../../lib/jones-v315.js';
 import * as goldAbeV316 from '../../lib/gold-abe-v316.js';
 import * as cdiSeverityV317 from '../../lib/cdi-severity-v317.js';
+import * as laEsophagitisV318 from '../../lib/la-esophagitis-v318.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -533,6 +534,7 @@ const MODULES = {
   'jones-v315.js': jonesV315,
   'gold-abe-v316.js': goldAbeV316,
   'cdi-severity-v317.js': cdiSeverityV317,
+  'la-esophagitis-v318.js': laEsophagitisV318,
   'tb-testing.js': tbTesting,
 };